corrected progress bar to fix #2
added a new counter that not only count rows that have a lookup identifier or that have missing data fields, but that have both: rows that have missing datafields and the potential to be enriched because there is a lookup identifier. This new counter works across fields we want to lookup. Counters like this already existed for each field, but not across.